fix: add null-guards to dialog close callbacks in savings, clients, and loans views to prevent runtime errors

+++ src/app/savings/savings-account-view/savings-account-view.component.ts
@@ -281,6 +281,9 @@ export class SavingsAccountViewComponent implements OnInit {
       data: { deleteContext: `savings account with id: ${this.savingsAccountData.id}` }
     });
     deleteSavingsAccountDialogRef.afterClosed().subscribe((response: any) => {
+      if (!response) {
+        return;
+      }
       if (response.delete) {
         this.savingsService.deleteSavingsAccount(this.savingsAccountData.id).subscribe(() => {
           this.router.navigate(['../../'], { relativeTo: this.route });
@@ -295,6 +298,9 @@ export class SavingsAccountViewComponent implements OnInit {
   private calculateInterest() {
     const calculateInterestAccountDialogRef = this.dialog.open(CalculateInterestDialogComponent);
     calculateInterestAccountDialogRef.afterClosed().subscribe((response: any) => {
+      if (!response) {
+        return;
+      }
       if (response.confirm) {
         this.savingsService
           .executeSavingsAccountCommand(this.savingsAccountData.id, 'calculateInterest', {})
@@ -311,6 +317,9 @@ export class SavingsAccountViewComponent implements OnInit {
   private postInterest() {
     const postInterestAccountDialogRef = this.dialog.open(PostInterestDialogComponent);
     postInterestAccountDialogRef.afterClosed().subscribe((response: any) => {
+      if (!response) {
+        return;
+      }
       if (response.confirm) {
         this.savingsService
           .executeSavingsAccountCommand(this.savingsAccountData.id, 'postInterest', {})
@@ -329,6 +338,9 @@ export class SavingsAccountViewComponent implements OnInit {
       data: { isEnable: true }
     });
     deleteSavingsAccountDialogRef.afterClosed().subscribe((response: any) => {
+      if (!response) {
+        return;
+      }
       if (response.confirm) {
         this.savingsService
           .executeSavingsAccountUpdateCommand(this.savingsAccountData.id, 'updateWithHoldTax', { withHoldTax: true })
@@ -347,6 +359,9 @@ export class SavingsAccountViewComponent implements OnInit {
       data: { isEnable: false }
     });
     disableWithHoldTaxDialogRef.afterClosed().subscribe((response: any) => {
+      if (!response) {
+        return;
+      }
       if (response.confirm) {
         this.savingsService
           .executeSavingsAccountUpdateCommand(this.savingsAccountData.id, 'updateWithHoldTax', { withHoldTax: false })
@@ -378,6 +393,9 @@ export class SavingsAccountViewComponent implements OnInit {
       command = 'unblockDebit';
     }
     unblockSavingsAccountDialogRef.afterClosed().subscribe((response: { confirm: any }) => {
+      if (!response) {
+        return;
+      }
       if (response.confirm) {
         this.savingsService.executeSavingsAccountCommand(this.savingsAccountData.id, command, {}).subscribe(() => {
           this.reload();
